Add trailing semicolons for toShell
tcsh complains if we try to call setenv three times in the same command
(for PATH, MANPATH, and INFOPATH).  Adding semicolons between them
causes them to be separate commands.  Since it is generally cleaner to
have the variables set in separate commands, use the semicolon always,
for all shells.

Bump the version to 10.4.2

Signed-off-by: Jonathan Reed <jdreed@mit.edu>